PL-204 - Added column wrappers which are used by the evaluation logic
Summary:
These are used in the expression evaluator (and soon to be changed UDF wrapper code).
They provide a type-erased means of accessing an std::vector so we can use them in expression evaluator.

Test Plan: Added unit tests.

Reviewers: #engineering, michelle

Reviewed By: #engineering, michelle

Differential Revision: https://phab.corp.pixielabs.ai/D231

GitOrigin-RevId: a090573ae08a3da2df399ee0848b9fd0b1ee6ae4

+++ src/carnot/udf/udf.h
@@ -160,7 +160,7 @@ struct UDFDataTypeTraits<udf::UDFDataType::FLOAT64> {
   using arrow_type = arrow::DoubleType;
   using arrow_builder_type = arrow::DoubleBuilder;
   using arrow_array_type = arrow::DoubleArray;
-  static constexpr arrow::Type::type arrow_type_id = arrow::Type::FLOAT;
+  static constexpr arrow::Type::type arrow_type_id = arrow::Type::DOUBLE;
 };
 
 template <>
